+ validations

+++ src/main/java/com/awbd/awbd/service/VehicleService.java
@@ -1,24 +1,20 @@
 package com.awbd.awbd.service;
 
 import com.awbd.awbd.config.SecurityUtil;
-import com.awbd.awbd.dto.VehicleCreationDto;
 import com.awbd.awbd.dto.VehicleDto;
+import com.awbd.awbd.entity.Appointment;
 import com.awbd.awbd.entity.Client;
-import com.awbd.awbd.entity.User;
 import com.awbd.awbd.entity.Vehicle;
 import com.awbd.awbd.mapper.VehicleMapper;
+import com.awbd.awbd.repository.AppointmentRepository;
 import com.awbd.awbd.repository.ClientRepository;
-import com.awbd.awbd.repository.UserRepository;
 import com.awbd.awbd.repository.VehicleRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -27,38 +23,34 @@ public class VehicleService {
     private final VehicleRepository vehicleRepository;
     private final VehicleMapper vehicleMapper;
     private final ClientRepository clientRepository;
-
-    public List<VehicleDto> findAll() {
-        return vehicleRepository.findAll().stream()
-                .map(vehicleMapper::toVehicleDto)
-                .toList();
-    }
+    private final AppointmentRepository appointmentRepository;
 
     @Transactional
     public void save(VehicleDto vehicleDto) {
         String username = SecurityUtil.getSessionUsername();
         Client client = clientRepository.findByUsername(username);
 
-        Vehicle vehicle;
-
-        if (vehicleDto.getId() != null) {
-            vehicle = vehicleRepository.findById(vehicleDto.getId())
-                    .orElseThrow(() -> new RuntimeException("Vehicle not found."));
-            vehicleMapper.updateVehicleFromDto(vehicleDto, vehicle);
-        } else {
-            vehicle = vehicleMapper.toVehicle(vehicleDto, client);
-        }
+        Vehicle vehicle = vehicleMapper.toVehicle(vehicleDto, client);
         vehicleRepository.save(vehicle);
     }
 
     public VehicleDto findById(Long id) {
-        Vehicle vehicle = vehicleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Vehicle not found."));
+        Vehicle vehicle = vehicleRepository.findVehicleById(id);
         return vehicleMapper.toVehicleDto(vehicle);
     }
 
     @Transactional
     public void deleteById(Long id) {
+        List<Appointment> appointments = appointmentRepository.findByVehicle_Id(id);
+        if (appointments.stream().anyMatch(appointment -> appointment.getReceipt() == null)){
+            throw new RuntimeException("Vehicle is used in an unfinished appointment.");
+        }
+
+        for (Appointment appointment : appointments) {
+            appointment.setVehicle(null);
+        }
+
+        appointmentRepository.saveAll(appointments);
         vehicleRepository.deleteById(id);
     }
 
@@ -70,5 +62,11 @@ public class VehicleService {
                 .stream()
                 .map(vehicleMapper::toVehicleDto)
                 .toList();
+    }
+
+    public void ensureNotInUse(Long id) {
+        if (appointmentRepository.existsByVehicle_Id(id)) {
+            throw new RuntimeException("Vehicle is used in an unfinished appointment.");
+        }
     }
 }
